Move OpCode.java, Performable.java and Regular.java to testclasses (#2163)
